refactor: Rename index files to YouTubePlaylist
