+++ src/YouTubePlaylist.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import YouTubePlaylist from "./index";
 import { test } from "@jest/globals";
+import YouTubePlaylist from "./YouTubePlaylist";
 
 test("youtube playlist renders correctly", () => {
   render(<YouTubePlaylist apiKey="" playlistId="" uniqueName="" />);
